requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
psycopg2-binary

